+++ source/renderer/app/api/staking/types.js
@@ -21,6 +21,14 @@ export type AdaApiStakePool = {
     },
   },
   apparent_performance: number,
+  cost: {
+    quantity: number,
+    unit: 'lovelace',
+  },
+  margin: {
+    quantity: number,
+    unit: 'percent',
+  },
   metadata: {
     owner: string,
     ticker: string, // [3 .. 5] characters
@@ -74,4 +82,10 @@ export type DelegationFee = {
     quantity: number,
     unit: WalletUnits.LOVELACE,
   },
+};
+
+export type QuitStakePoolRequest = {
+  stakePoolId: string,
+  walletId: string,
+  passphrase: string,
 };