+++ source/renderer/app/components/staking/stake-pools/StakePoolTooltip.js
@@ -1,7 +1,6 @@
 // @flow
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
-import BigNumber from 'bignumber.js';
 import { defineMessages, intlShape, FormattedMessage } from 'react-intl';
 import { Button } from 'react-polymorph/lib/components/Button';
 import classnames from 'classnames';
@@ -15,7 +14,6 @@ import closeCross from '../../../assets/images/close-cross.inline.svg';
 import externalLinkIcon from '../../../assets/images/link-ic.inline.svg';
 import { getColorFromRange } from '../../../utils/colors';
 import { formattedWalletAmount, shortNumber } from '../../../utils/formatters';
-import { LOVELACES_PER_ADA } from '../../../config/numbersConfig';
 import { rangeMap } from '../../../utils/rangeMap';
 import {
   THUMBNAIL_HEIGHT,
@@ -44,6 +42,11 @@ const messages = defineMessages({
     defaultMessage: '!!!Profit margin:',
     description: '"Profit margin" for the Stake Pools Tooltip page.',
   },
+  costPerEpoch: {
+    id: 'staking.stakePools.tooltip.costPerEpoch',
+    defaultMessage: '!!!Cost per epoch:',
+    description: '"Cost per epoch" for the Stake Pools Tooltip page.',
+  },
   performance: {
     id: 'staking.stakePools.tooltip.performance',
     defaultMessage: '!!!Performance:',
@@ -60,9 +63,9 @@ const messages = defineMessages({
     description: '"Retirement" for the Stake Pools Tooltip page.',
   },
   // cost: {
-  //   id: 'staking.stakePools.tooltip.cost',
-  //   defaultMessage: '!!!Cost:',
-  //   description: 'Cost" for the Stake Pools Tooltip page.',
+  //  id: 'staking.stakePools.tooltip.cost',
+  //  defaultMessage: '!!!Operating Costs:',
+  //  description: 'Cost" for the Stake Pools Tooltip page.',
   // },
   // pledge: {
   //   id: 'staking.stakePools.tooltip.pledge',
@@ -384,6 +387,8 @@ export default class StakePoolTooltip extends Component<Props, State> {
       producedBlocks,
       retiring,
       pledgeAddress,
+      cost,
+      profitMargin,
     } = stakePool;
 
     const componentClassnames = classnames([
@@ -426,13 +431,17 @@ export default class StakePoolTooltip extends Component<Props, State> {
             </div>
           )}
           <div className={styles.description}>{description}</div>
-          <button
-            className={styles.homepage}
+
+          <span
+            role="presentation"
+            aria-hidden
+            className={styles.homepageAddress}
             onClick={() => onOpenExternalLink(homepage)}
           >
-            <span className={styles.homepageContent}>{homepage}</span>
+            {homepage}
             <SVGInline svg={externalLinkIcon} />
-          </button>
+          </span>
+
           <dl className={styles.table}>
             <dt>{intl.formatMessage(messages.ranking)}</dt>
             <dd className={styles.ranking}>
@@ -446,28 +455,34 @@ export default class StakePoolTooltip extends Component<Props, State> {
             </dd>
             <dt>{intl.formatMessage(messages.controlledStake)}</dt>
             <dd className={styles.defaultColor}>
-              <span>
-                {formattedWalletAmount(
-                  new BigNumber(controlledStake).dividedBy(LOVELACES_PER_ADA),
-                  true,
-                  false
-                )}
-              </span>
+              <span>{formattedWalletAmount(controlledStake, true, false)}</span>
             </dd>
-            {/* <dt>{intl.formatMessage(messages.profitMargin)}</dt>
+            <dt>{intl.formatMessage(messages.profitMargin)}</dt>
             <dd className={styles.profitMargin}>
               <span
                 style={{
                   background: getColorFromRange(profitMargin, {
                     darken,
                     alpha,
-                    reverse,
                   }),
                 }}
               >
-                {parseFloat(profitMargin.toFixed(2))}%
+                {`${parseFloat(profitMargin.toFixed(2))}%`}
               </span>
-            </dd> */}
+            </dd>
+            <dt>{intl.formatMessage(messages.costPerEpoch)}</dt>
+            <dd className={styles.profitMargin}>
+              <span
+                style={{
+                  background: getColorFromRange(profitMargin, {
+                    darken,
+                    alpha,
+                  }),
+                }}
+              >
+                {`${formattedWalletAmount(cost, true, false)}`}
+              </span>
+            </dd>
             <dt>{intl.formatMessage(messages.performance)}</dt>
             <dd className={styles.performance}>
               <span
@@ -486,22 +501,22 @@ export default class StakePoolTooltip extends Component<Props, State> {
             <dd className={styles.defaultColor}>
               <span>{shortNumber(producedBlocks)}</span>
             </dd>
-            {/*
-            <dt>{intl.formatMessage(messages.cost)}</dt>
+            {/* <dt>{intl.formatMessage(messages.cost)}</dt>
             <dd>
               <span
                 style={{
-                  background: getColorFromRange(cost, {
+                  background: getColorFromRange(shortNumber(cost), {
                     darken,
                     alpha,
                   }),
                 }}
               >
-                {formattedWalletAmount(cost)}
+                {formattedWalletAmount(shortNumber(cost))}
               </span>
-            </dd></dl>
-            <dt>{intl.formatMessage(messages.pledge)}</dt>
-            <dd>
+            </dd> */}
+          </dl>
+          {/* <dt>{intl.formatMessage(messages.pledge)}</dt> */}
+          {/* <dd>
               <span
                 style={{
                   background: getColorFromRange(pledge, {
@@ -512,16 +527,14 @@ export default class StakePoolTooltip extends Component<Props, State> {
               >
                 {formattedWalletAmount(pledge)}
               </span>
-            </dd>
-            */}
-          </dl>
+            </dd> */}
           <button
-            className={styles.homepage}
+            className={styles.pledgeAddress}
             onClick={() =>
               onOpenExternalLink(getPledgeAddressUrl(pledgeAddress))
             }
           >
-            <span className={styles.homepageContent}>
+            <span className={styles.pledgeAddressContent}>
               {intl.formatMessage(messages.pledgeAddressLabel)}
             </span>
             <SVGInline svg={externalLinkIcon} />