+++ source/renderer/app/domains/StakePool.js
@@ -21,13 +21,12 @@ export type StakePoolProps = {
   performance: number,
   controlledStake: BigNumber,
   producedBlocks: number,
-  createdAt: Date,
   description: string,
-  // cost: BigNumber,
+  cost: BigNumber,
   // pledge: BigNumber,
   isCharity: boolean,
   name: string,
-  // profitMargin: number,
+  profitMargin: number,
   ranking: number,
   retiring?: ?Date,
 };
@@ -39,10 +38,9 @@ export default class StakePool {
   @observable pledgeAddress: string;
   @observable performance: number;
   @observable producedBlocks: number;
-  @observable controlledStake: number;
+  @observable controlledStake: BigNumber;
   // @observable pledge: BigNumber;
-  // @observable cost: BigNumber;
-  @observable createdAt: Date;
+  @observable cost: BigNumber;
   @observable description: string = '';
   @observable isCharity: boolean;
   @observable name: string = '';